f1-edge/requirements.txt
f1-23-telemetry
flask
flask-socketio
paho-mqtt
eventlet